+++ clevertap.js
@@ -4090,6 +4090,15 @@
   };
   const initializeWebInbox = logger => {
     return new Promise((resolve, reject) => {
+      // Adding this as a band-aid for SUC-126380
+      // Adds ct-web-inbox element in dom which is not visible if Web Inbox Config in LS
+      document.addEventListener('readystatechange', function () {
+        if (document.readyState === 'complete') {
+          addWebInbox(logger);
+          resolve();
+        }
+      });
+
       if (document.readyState === 'complete') {
         addWebInbox(logger);
         resolve();
@@ -4366,7 +4375,7 @@
 
     if (search === '?ctBuilderSDKCheck') {
       if (parentWindow) {
-        const sdkVersion = '1.11.11';
+        const sdkVersion = '1.11.14';
         parentWindow.postMessage({
           message: 'SDKVersion',
           accountId,
@@ -5359,9 +5368,10 @@
     enable() {
       let options = arguments.length > 0 && arguments[0] !== undefined ? arguments[0] : {};
       const {
-        swPath
+        swPath,
+        skipDialog
       } = options;
-      enablePush(_classPrivateFieldLooseBase(this, _logger$5)[_logger$5], _classPrivateFieldLooseBase(this, _account$2)[_account$2], _classPrivateFieldLooseBase(this, _request$4)[_request$4], swPath);
+      enablePush(_classPrivateFieldLooseBase(this, _logger$5)[_logger$5], _classPrivateFieldLooseBase(this, _account$2)[_account$2], _classPrivateFieldLooseBase(this, _request$4)[_request$4], swPath, skipDialog);
     }
 
     _processOldValues() {
@@ -5798,7 +5808,7 @@
       updatePushConfig();
     }
   };
-  const enablePush = (logger, account, request, customSwPath) => {
+  const enablePush = (logger, account, request, customSwPath, skipDialog) => {
     const _pushConfig = StorageManager.readFromLSorCookie(WEBPUSH_CONFIG) || {};
 
     $ct.pushConfig = _pushConfig;
@@ -5818,6 +5828,13 @@
       request,
       account
     });
+
+    if (skipDialog) {
+      notificationHandler.setApplicationServerKey(appServerKey);
+      notificationHandler.setUpWebPushNotifications(null, swPath, null, null);
+      return;
+    }
+
     const {
       showBox,
       boxType,
@@ -5905,14 +5922,28 @@
     wrapper.appendChild(pnCard);
     wrapper.appendChild(overlayDiv);
     setElementPosition(pnCard, style.card.position);
+
+    if (!configData.isPreview) {
+      if ('Notification' in window && Notification !== null) {
+        if (Notification.permission === 'granted') {
+          notificationHandler.setApplicationServerKey(appServerKey);
+          notificationHandler.setUpWebPushNotifications(null, swPath, null, null);
+          return;
+        } else if (Notification.permission === 'denied') {
+          return;
+        }
+      }
+    }
+
     const now = new Date().getTime() / 1000;
     const lastNotifTime = StorageManager.getMetaProp('webpush_last_notif_time');
-    const popupFrequency = content.popupFrequency || 7 * 24 * 60 * 60;
+    const popupFrequency = content.popupFrequency || 7; // number of days
 
     if (!lastNotifTime || now - lastNotifTime >= popupFrequency * 24 * 60 * 60) {
       document.body.appendChild(wrapper);
 
       if (!configData.isPreview) {
+        StorageManager.setMetaProp('webpush_last_notif_time', now);
         addEventListeners(wrapper);
       }
     }
@@ -5991,7 +6022,6 @@
       notificationHandler.setUpWebPushNotifications(null, swPath, null, null);
     });
     secondaryButton.addEventListener('click', () => {
-      StorageManager.setMetaProp('webpush_last_notif_time', Date.now() / 1000);
       removeWrapper();
     });
   };
@@ -7475,7 +7505,7 @@
       let proto = document.location.protocol;
       proto = proto.replace(':', '');
       dataObject.af = { ...dataObject.af,
-        lib: 'web-sdk-v1.11.11',
+        lib: 'web-sdk-v1.11.14',
         protocol: proto,
         ...$ct.flutterVersion
       }; // app fields
@@ -8486,10 +8516,12 @@
 
 
       this.getInboxMessageUnreadCount = () => {
-        if ($ct.inbox) {
-          return $ct.inbox.unviewedCounter;
-        } else {
-          _classPrivateFieldLooseBase(this, _logger$a)[_logger$a].debug('No unread messages');
+        try {
+          const unreadMessages = this.getUnreadInboxMessages();
+          const result = Object.keys(unreadMessages).length;
+          return result;
+        } catch (e) {
+          _classPrivateFieldLooseBase(this, _logger$a)[_logger$a].error('Error in getInboxMessageUnreadCount' + e);
         }
       }; // Get All Inbox messages
 
@@ -8500,10 +8532,21 @@
 
 
       this.getUnreadInboxMessages = () => {
-        if ($ct.inbox) {
-          return $ct.inbox.unviewedMessages;
-        } else {
-          _classPrivateFieldLooseBase(this, _logger$a)[_logger$a].debug('No unread messages');
+        try {
+          const messages = getInboxMessages();
+          const result = {};
+
+          if (Object.keys(messages).length > 0) {
+            for (const message in messages) {
+              if (messages[message].viewed === 0) {
+                result[message] = messages[message];
+              }
+            }
+          }
+
+          return result;
+        } catch (e) {
+          _classPrivateFieldLooseBase(this, _logger$a)[_logger$a].error('Error in getUnreadInboxMessages' + e);
         }
       }; // Get message object belonging to the given message id only. Message id should be a String
 
@@ -8556,10 +8599,13 @@
 
 
       this.markReadInboxMessage = messageId => {
-        const unreadMsg = $ct.inbox.unviewedMessages;
         const messages = getInboxMessages();
 
-        if ((messageId !== null || messageId !== '') && unreadMsg.hasOwnProperty(messageId)) {
+        if ((messageId !== null || messageId !== '') && messages.hasOwnProperty(messageId)) {
+          if (messages[messageId].viewed === 1) {
+            return _classPrivateFieldLooseBase(this, _logger$a)[_logger$a].error('Message already viewed' + messageId);
+          }
+
           const ctInbox = document.querySelector('ct-web-inbox');
 
           if (ctInbox) {
@@ -8607,8 +8653,8 @@
 
 
       this.markReadAllInboxMessage = () => {
-        const unreadMsg = $ct.inbox.unviewedMessages;
         const messages = getInboxMessages();
+        const unreadMsg = this.getUnreadInboxMessages();
 
         if (Object.keys(unreadMsg).length > 0) {
           const msgIds = Object.keys(unreadMsg);
@@ -9167,7 +9213,7 @@
     }
 
     getSDKVersion() {
-      return 'web-sdk-v1.11.11';
+      return 'web-sdk-v1.11.14';
     }
 
     defineVariable(name, defaultValue) {